requirements.txt
numpy
scipy
torch
torchvision
Pillow
scikit-image
joblib
scikit-learn
flask
